+++ espnff/espnff.py
@@ -1,6 +1,22 @@
 import requests
 
-from .utils import two_step_dominance, power_points
+from .utils import two_step_dominance, power_points, deprecated_property
+
+
+class ESPNFFException(Exception):
+    pass
+
+
+class PrivateLeagueException(ESPNFFException):
+    pass
+
+
+class InvalidLeagueException(ESPNFFException):
+    pass
+
+
+class UnknownLeagueException(ESPNFFException):
+    pass
 
 
 class League(object):
@@ -10,16 +26,33 @@ class League(object):
         self.year = year
         self.ENDPOINT = "http://games.espn.com/ffl/api/v2/"
         self.teams = []
-        self._fetch_teams()
+        self._fetch_league()
 
     def __repr__(self):
         return 'League %s, %s Season' % (self.league_id, self.year)
 
-    def _fetch_teams(self):
-        '''Fetch teams in league'''
-        url = "%sleagueSettings?leagueId=%s&seasonId=%s"
-        r = requests.get(url % (self.ENDPOINT, self.league_id, self.year))
+    def _fetch_league(self):
+        params = {
+            'leagueId': self.league_id,
+            'seasonId': self.year
+        }
+        r = requests.get('%sleagueSettings' % (self.ENDPOINT, ), params=params)
+        self.status = r.status_code
         data = r.json()
+
+        if self.status == 401:
+            raise PrivateLeagueException(data['error'][0]['message'])
+
+        elif self.status == 404:
+            raise InvalidLeagueException(data['error'][0]['message'])
+
+        elif self.status != 200:
+            raise UnknownLeagueException('Unknown %s Error' % self.status)
+
+        self._fetch_teams(data)
+
+    def _fetch_teams(self, data):
+        '''Fetch teams in league'''
         teams = data['leaguesettings']['teams']
 
         for team in teams:
@@ -29,7 +62,7 @@ class League(object):
         for team in self.teams:
             for week, matchup in enumerate(team.schedule):
                 for opponent in self.teams:
-                    if matchup == opponent.teamId:
+                    if matchup == opponent.team_id:
                         team.schedule[week] = opponent
 
         # calculate margin of victory
@@ -39,20 +72,20 @@ class League(object):
                 team.mov.append(mov)
 
         # sort by team ID
-        self.teams = sorted(self.teams, key=lambda x: x.teamId, reverse=False)
+        self.teams = sorted(self.teams, key=lambda x: x.team_id, reverse=False)
 
     def power_rankings(self, week):
         '''Return power rankings for any week'''
 
         # calculate win for every week
         win_matrix = []
-        teams_sorted = sorted(self.teams, key=lambda x: x.teamId,
+        teams_sorted = sorted(self.teams, key=lambda x: x.team_id,
                               reverse=False)
 
         for team in teams_sorted:
             wins = [0]*32
             for mov, opponent in zip(team.mov[:week], team.schedule[:week]):
-                opp = int(opponent.teamId)-1
+                opp = int(opponent.team_id)-1
                 if mov > 0:
                     wins[opp] += 1
             win_matrix.append(wins)
@@ -64,15 +97,15 @@ class League(object):
 class Team(object):
     '''Teams are part of the league'''
     def __init__(self, data):
-        self.teamId = data['teamId']
-        self.teamAbbrev = data['teamAbbrev']
-        self.teamName = "%s %s" % (data['teamLocation'], data['teamNickname'])
-        self.divisionId = data['division']['divisionId']
-        self.divisionName = data['division']['divisionName']
+        self.team_id = data['teamId']
+        self.team_abbrev = data['teamAbbrev']
+        self.team_name = "%s %s" % (data['teamLocation'], data['teamNickname'])
+        self.division_id = data['division']['divisionId']
+        self.division_name = data['division']['divisionName']
         self.wins = data['record']['overallWins']
         self.losses = data['record']['overallLosses']
-        self.pointsFor = data['record']['pointsFor']
-        self.pointsAgainst = data['record']['pointsAgainst']
+        self.points_for = data['record']['pointsFor']
+        self.points_against = data['record']['pointsAgainst']
         self.owner = "%s %s" % (data['owners'][0]['firstName'],
                                 data['owners'][0]['lastName'])
         self.schedule = []
@@ -81,15 +114,23 @@ class Team(object):
         self._fetch_schedule(data)
 
     def __repr__(self):
-        return 'Team %s' % self.teamName
+        return 'Team %s' % self.team_name
+
+    teamId = deprecated_property('teamId', 'team_id')
+    teamAbbrev = deprecated_property('teamAbbrev', 'team_abbrev')
+    teamName = deprecated_property('teamName', 'team_name')
+    divisionId = deprecated_property('divisionId', 'division_id')
+    divisionName = deprecated_property('divisionName', 'division_name')
+    pointsFor = deprecated_property('pointsFor', 'points_for')
+    pointsAgainst = deprecated_property('pointsAgainst', 'points_against')
 
     def _fetch_schedule(self, data):
         '''Fetch schedule and scores for team'''
         matchups = data['scheduleItems']
 
         for matchup in matchups:
-            if matchup['matchups'][0]['isBye'] is False:
-                if matchup['matchups'][0]['awayTeamId'] == self.teamId:
+            if not matchup['matchups'][0]['isBye']:
+                if matchup['matchups'][0]['awayTeamId'] == self.team_id:
                     score = matchup['matchups'][0]['awayTeamScores'][0]
                     opponentId = matchup['matchups'][0]['homeTeamId']
                 else: